LV2: Convert to NodeProvider & Fix Problems Scanning LV2
Several LV2 fixes and improvements. Closes #643

* lv2: Bind control ports to plugins
* lv2 ui: resolve X11 "BadWindow" errors
* lv2: less glitchy plugin window instantiation.
* lv2: better, not perfect, UI shutdown order
* juce: add gui_extra.hpp
* force usage of std::placeholders
* Remove deploy option from LVTK
* vtable: fix missing symbols (maybe)
* lv2: bring back UI support as it was in 0.46
* Add stubs for checking Procesor editor
* lv2: much better performance loading
* Work out lv2 instance
* lv2: take lvtk::World features when instantating
* Uset FeatureList in plugin Model
* lv2: integrating LVTK host, WIP
* updater: make launchable from preferences
* windows: use component script to install element.exe
* lvtk: ensure to use static build for windows.
* lv2: clean out old plugin scanner
* lv2: update lvtk.wrap to use v3/host branch
* linux: fix problems with plugin scanner crashing.
* lv2: scan with plugin scanner worker.
* resolve warnings. clean up logging and license info
* portbuffer: compile new code with meson
* vscode: update linux paths
* lv2: use brute force to "scan" LV2 for re-implementation.
* Linux: resolve GCC warnings.
* vscode: update linux include path
* signtool: add win32 codesigning. Untested.

+++ src/gui/GraphEditorComponent.cpp
@@ -700,8 +700,8 @@ ConnectorComponent* GraphEditorComponent::getComponentForConnection (const Arc& 
         if (ConnectorComponent* const c = dynamic_cast<ConnectorComponent*> (getChildComponent (i)))
             if (c->sourceFilterID == arc.sourceNode
                 && c->destFilterID == arc.destNode
-                && c->sourceFilterChannel == arc.sourcePort
-                && c->destFilterChannel == arc.destPort)
+                && (uint32_t) c->sourceFilterChannel == arc.sourcePort
+                && (uint32_t) c->destFilterChannel == arc.destPort)
                 return c;
     }
